Some cleaning

+++ src/main/kotlin/dev/axion/types/impl/BooleanWasmType.kt
@@ -4,5 +4,5 @@ import dev.axion.types.WasmType
 
 class BooleanWasmType(private val boolean: Boolean) : WasmType(
     boolean,
-    toWasmerValue = { (if(boolean) 1 else 0) as Integer },
+    toWasmerValue = { if(boolean) 1 else 0 },
 )
