+++ src/main/kotlin/dev/axion/types/impl/IntegerWasmType.kt
@@ -4,5 +4,5 @@ import dev.axion.types.WasmType
 
 class IntegerWasmType(private val int: Int) : WasmType(
     int,
-    toWasmerValue = { int as Integer },
+    toWasmerValue = { int },
 )
