+++ src/main/kotlin/dev/axion/types/impl/LongWasmType.kt
@@ -4,5 +4,5 @@ import dev.axion.types.WasmType
 
 class LongWasmType(private val long: Long) : WasmType(
     long,
-    toWasmerValue = { long as java.lang.Long },
+    toWasmerValue = { long },
 )
