+++ src/main/kotlin/dev/axion/types/impl/ShortWasmType.kt
@@ -4,5 +4,5 @@ import dev.axion.types.WasmType
 
 class ShortWasmType(private val short: Short) : WasmType(
     short,
-    toWasmerValue = { short.toInt() as Integer }
+    toWasmerValue = { short.toInt() }
 )
